Minor: Renaming CouncilGroupSearch files to the default format
